+++ setup.py
@@ -21,6 +21,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
+from __future__ import absolute_import
 import os
 import errno
 import sys
@@ -31,6 +32,8 @@ from setuptools.command.test import test as TestCommand
 
 import re
 
+from utils import fs
+from utils import doc
 
 ## CONFIG
 target_version = '2.1.1'
@@ -56,52 +59,6 @@ if is_dev_version:
     print("Target: %s\n" % target_version)
 else:
     print("!!!>>> This is a RELEASE version <<<!!!\n")
-
-## Filesystem utilities
-def read_file(path):
-    """Reads file into string."""
-    with open(path, 'r') as f:
-        data = f.read()
-    return data
-
-
-def mkdir_p(path):
-    """Like `mkdir -p` in unix"""
-    if not path.strip():
-        return
-    try:
-        os.makedirs(path)
-    except OSError as e:
-        if e.errno == errno.EEXIST and os.path.isdir(path):
-            pass
-        else:
-            raise
-
-
-def rm_f(path):
-    """Like `rm -f` in unix"""
-    try:
-        os.unlink(path)
-    except OSError as e:
-        if e.errno == errno.ENOENT:
-            pass
-        else:
-            raise
-
-
-def update_file(path, data):
-    """Writes data to path, creating path if it doesn't exist"""
-    # delete file if already exists
-    rm_f(path)
-
-    # create parent dirs if needed
-    parent_dir = os.path.dirname(path)
-    if not os.path.isdir(os.path.dirname(parent_dir)):
-        mkdir_p(parent_dir)
-
-    # write file
-    with open(path, 'w') as f:
-        f.write(data)
 
 
 ## PyTest
@@ -139,21 +96,21 @@ def get_rst():
 ## Convert README to rst for PyPI
 def generate_rst():
     """Converts Markdown to RST for PyPI"""
-    md = read_file("README.md")
+    md = fs.read_file("README.md")
 
     md = pypi_sanitize_markdown(md)
     rst = markdown2rst(md)
     rst = pypi_prepare_rst(rst)
 
-    changes_md = pypi_sanitize_markdown(read_file("CHANGELOG.md"))
+    changes_md = pypi_sanitize_markdown(fs.read_file("CHANGELOG.md"))
     changes_rst = markdown2rst(changes_md)
     rst += "\n" + changes_rst
 
     # Write it
     if create_rst:
-        update_file('README.rst', rst)
+        fs.update_file('README.rst', rst)
     else:
-        rm_f('README.rst')
+        fs.rm_f('README.rst')
 
     return rst
